Downloading 1GB on loopback on Windows now achieves 963Mbps with BBR, 971 Mbps with BBR.
On Ubuntu VM, I get 1213 Mbps with BBR, 1361 Mbps with Cubic. I.e., 30% faster.